% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_per_mean.R
\name{actual.within.shift.var.per.mean.loop}
\alias{actual.within.shift.var.per.mean.loop}
\title{Convert all within shift actual staffing variances}
\usage{
actual.within.shift.var.per.mean.loop(data, grid)
}
\arguments{
\item{data}{}

\item{grid}{}
}
\description{
Converts from varaiance to variance per mean.
}
