% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_per_mean.R
\name{actual.within.shift.var.per.mean}
\alias{actual.within.shift.var.per.mean}
\title{Convert within shift variance for actual staffing levels.}
\usage{
actual.within.shift.var.per.mean(
  data,
  window,
  staff = "Registered",
  type = "Count"
)
}
\arguments{
\item{data}{Data frame}

\item{window}{A lagging period present in 'data'}

\item{staff}{A staff type in 'data'}

\item{type}{A staffing metric in 'data'}
}
\description{
Convert within shift variance for actual staffing levels.
}
