% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Acuity.Data.R
\name{acuity.score.f}
\alias{acuity.score.f}
\title{Calculate a weighted acuity score using the SNCT multipliers.}
\usage{
acuity.score.f(L0, L1A, L1B, L2, L3)
}
\description{
Calculate a weighted acuity score using the SNCT multipliers.
}
