% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_per_mean.R
\name{acuity.var.features}
\alias{acuity.var.features}
\title{Convert all raw planned staffing variances}
\usage{
acuity.var.features(data, grid)
}
\arguments{
\item{data}{}

\item{grid}{}
}
\description{
Converts from varaiance to variance per mean.
}
