% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Acuity.Data.R
\name{acuity_score_f}
\alias{acuity_score_f}
\title{Calculate a weighted acuity score using the SNCT multipliers.}
\usage{
acuity_score_f(l0, l1a, l1b, l2, l3)
}
\description{
Calculate a weighted acuity score using the SNCT multipliers.
}
