% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.bed.size.R
\name{add.bed.size}
\alias{add.bed.size}
\title{Add 'bed size' to data set}
\usage{
add.bed.size(data, date.column, ward)
}
\arguments{
\item{data}{A data frame to add column to}

\item{date.column}{Name of the 'Date' reference column in data}

\item{ward}{Name of a ward (string)}
}
\description{
Add a column listing number of beds a ward has to a data set - 
depending on ward and date.
}
