% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly_occupancy.R
\name{add.days.occupied}
\alias{add.days.occupied}
\title{Append 'pas.days.occupied' data set to another data set}
\usage{
add.days.occupied(data, date.column)
}
\arguments{
\item{data}{A data frame with 'Ward' and 'date.column' fields}

\item{date.column}{A field of 'data' that is a data/ time}
}
\description{
Merge on Ward, Year and Month
}
