% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly_average_establishment.R
\name{add.establishment}
\alias{add.establishment}
\title{Add 'Establishment' to a data set}
\usage{
add.establishment(data, date.column)
}
\arguments{
\item{data}{Data frame to add est to.}

\item{date.column}{Column of 'data' to be used as a date}
}
\description{
Aligns on 'date.column' and 'Ward'
}
