% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.bed.size.R
\name{add_bed_size_ws}
\alias{add_bed_size_ws}
\title{Add 'bed size' to data set}
\usage{
add_bed_size_ws(data, date_column, ward)
}
\arguments{
\item{data}{A data frame to add column to}

\item{ward}{Name of a ward (string)}

\item{date.column}{Name of the 'Date' reference column in data}
}
\description{
Add a column listing number of beds a ward has to a data set -
depending on ward and date (where the data is a POSIXct object).
}
