% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{all.ward.month.merge}
\alias{all.ward.month.merge}
\title{Add exhaustive set of Wards and Dates to a data set}
\usage{
\method{all}{ward.month.merge}(data)
}
\arguments{
\item{data}{A data frame with Ward and Date fields.}
}
\value{
A data frame
}
\description{
Produces a data frame of all combinations of Wards and Dates in 'data',
and merges to 'data'.  Results in 'na' values where the Ward and Date
combination wasn't in 'data'
}
