% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sickness.R
\name{check.mend}
\alias{check.mend}
\title{Limit upper value}
\usage{
check.mend(month.end)
}
\arguments{
\item{month.end}{Date or date-time limit}
}
\description{
Make a function to limit a date to an upper value
}
