% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sickness.R
\name{check.mstart}
\alias{check.mstart}
\title{Limit lower value}
\usage{
check.mstart(month.start)
}
\arguments{
\item{month.start}{Date or date-time limit}
}
\description{
Make a function to limit a date to a lower value
}
