% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_work_demographics.R
\name{clean.demographics}
\alias{clean.demographics}
\title{Remap demographics in a data set based on "DemographicAnalysisGroups.xlsx"}
\usage{
clean.demographics(data, .list = demo.list)
}
\arguments{
\item{data}{A data frame}

\item{.list}{A mapped list of cardinality reduction.}
}
\description{
Remap demographics in a data set based on "DemographicAnalysisGroups.xlsx"
}
