% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete.R
\name{complete_f}
\alias{complete_f}
\title{Check which rows of a data frame are complete}
\usage{
complete_f(data)
}
\arguments{
\item{data}{A data frame}
}
\description{
Check which rows of a data frame are complete
}
