% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{credential.prompt}
\alias{credential.prompt}
\title{Generate DB connection object from config file.
Ask user for SQL credentials}
\usage{
credential.prompt()
}
\description{
Generate DB connection object from config file.
Ask user for SQL credentials
}
