% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_work_demographics.R
\name{encode.all.demos}
\alias{encode.all.demos}
\title{convert all given variables to a one hot encoding.}
\usage{
encode.all.demos(data, factor.cols)
}
\arguments{
\item{data}{A data frame}

\item{factor.cols}{A vector of column in data}
}
\description{
convert all given variables to a one hot encoding.
}
