% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_to_data.R
\name{esr_to_allocate}
\alias{esr_to_allocate}
\title{Map ESR locations to Allocate Wards}
\usage{
esr_to_allocate(values)
}
\arguments{
\item{values}{A vector of ESR locations}
}
\description{
Map ESR locations to Allocate Wards
}
