% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_work_demographics.R
\name{fetch.swd}
\alias{fetch.swd}
\title{Get shifts worked data for a given year, and mark where values are missing due to
scrambler}
\usage{
fetch.swd(year, .con = pkg_env$con)
}
\arguments{
\item{year}{Numeric year of data to convert}

\item{.con}{A DBI connection object}
}
\description{
Get shifts worked data for a given year, and mark where values are missing due to
scrambler
}
