% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate_shift_worked_functions.R
\name{fetch_sw}
\alias{fetch_sw}
\title{Get shifts workled data by year}
\usage{
fetch_sw(year)
}
\arguments{
\item{.con}{SQL engine connection}

\item{.year}{}
}
\description{
Get shifts workled data by year
}
