% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_work_demographics.R
\name{get.demographic.data}
\alias{get.demographic.data}
\title{Fetch and clean demographics data and cache to file.}
\usage{
get.demographic.data()
}
\description{
Fetch and clean demographics data and cache to file.
}
