% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/establishment.R
\name{get.establishment.by.year}
\alias{get.establishment.by.year}
\title{Generate monthly establihsment data sets}
\usage{
get.establishment.by.year(.year)
}
\arguments{
\item{.year}{Year of data to be processed}
}
\description{
Generate monthly establihsment data sets
}
