% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Leavers.R
\name{get.leavers}
\alias{get.leavers}
\title{Query list of leavers from 'staging' data base}
\usage{
get.leavers(sql.table = LEAVERS.SQL.TABLE)
}
\arguments{
\item{sql.table}{Name of data set in database}
}
\description{
Query list of leavers from 'staging' data base
}
