% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate_shift_worked_functions.R
\name{get_establishment}
\alias{get_establishment}
\title{Fetch ward establishments}
\usage{
get_establishment(.year)
}
\arguments{
\item{.year}{Year of data to be queried and processed}
}
\description{
Fetch ward establishments
}
