% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/establishment.R
\name{get_establishment_by_year}
\alias{get_establishment_by_year}
\title{Generate monthly establishment data sets}
\usage{
get_establishment_by_year(.year, .con = pkg_env$con)
}
\arguments{
\item{.year}{Year of data to be processed}

\item{.con}{A DBI connection object}
}
\description{
Generate monthly establishment data sets
}
