% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Leavers.R
\name{get_leavers}
\alias{get_leavers}
\title{Query list of leavers from 'staging' data base}
\usage{
get_leavers(sql_table = LEAVERS_SQL_TABLE, .con = pkg_env$con)
}
\arguments{
\item{sql_table}{Name of data set in database}
}
\description{
Query list of leavers from 'staging' data base
}
