% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria.R
\name{inc_exc_criteria}
\alias{inc_exc_criteria}
\title{Apply exclusion criteria to data set}
\usage{
inc_exc_criteria(data)
}
\arguments{
\item{data}{A data frame with columns 'Year' and 'Ward'}
}
\description{
Apply exclusion criteria to data set
}
