% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag.R.R
\name{lag.data.frame}
\alias{lag.data.frame}
\title{Produce a data set where observations have been lagged in time by a given value. 
If multiple lagging months are given - the data set produces duplicates.}
\usage{
lag.data.frame(data, date_col, months)
}
\arguments{
\item{data}{The data frame to be lagged}

\item{date_col}{The column of `data` to use for lagging}

\item{months}{[Integer] The number of months to lag data by.  Expects integer value(s)}
}
\description{
Produce a data set where observations have been lagged in time by a given value. 
If multiple lagging months are given - the data set produces duplicates.
}
