% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag.R.R
\name{lag_data_frame}
\alias{lag_data_frame}
\title{Produce a data set where observations have been lagged in time
  by a given value.
If multiple lagging months are given - the data set produces duplicates.}
\usage{
lag_data_frame(data, date_col, months)
}
\arguments{
\item{data}{The data frame to be lagged}

\item{date_col}{The column of `data` to use for lagging}

\item{months}{[Integer] The number of months to lag data by.
Expects integer value(s)}
}
\description{
Produce a data set where observations have been lagged in time
  by a given value.
If multiple lagging months are given - the data set produces duplicates.
}
