% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag.R.R
\name{lagged.group}
\alias{lagged.group}
\title{Convert data set into a lagged, grouped version.}
\usage{
lagged.group(data, date.col, window = 3:4, ...)
}
\arguments{
\item{data}{The data set to be converted NB: if no grouping variables given (`...`) 
data is expected to contain a column `Ward`.}

\item{date.col}{The column of  data to treat as the date for lagging.}

\item{window}{[optional] The number of months to lag the data by.}

\item{...}{[optional] Columns of data to group by.  If none given - 
defaults to `Ward` and the `Year`/`Month` produced by lagging `date.col`}
}
\description{
Convert data set into a lagged, grouped version.
}
