% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{lagged.process}
\alias{lagged.process}
\title{Apply a function to sets of lag windows}
\usage{
lagged.process(process.func)
}
\arguments{
\item{process.func}{A function expecting to receive a lag-window to produce data.}
}
\description{
Apply a function to sets of lag windows
}
