% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_to_data.R
\name{list_lookup}
\alias{list_lookup}
\title{Map a value to an entity in a list (returns "Unknown" is missing)}
\usage{
list_lookup(i, .list)
}
\arguments{
\item{i}{value to be mapped}

\item{.list}{Lookup list}
}
\description{
Map a value to an entity in a list (returns "Unknown" is missing)
}
