% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAS_data.R
\name{load.pas.data}
\alias{load.pas.data}
\title{Read pas ward stay data}
\usage{
load.pas.data(sql_table = "pas_wardstays")
}
\arguments{
\item{sql_table}{PAS table name in sql engine}
}
\description{
Read pas ward stay data
}
