% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAS_data.R
\name{make.admissions.discharges}
\alias{make.admissions.discharges}
\title{Calculate monthly admissions and discharges}
\usage{
make.admissions.discharges(sql.table = "pas_wardstays")
}
\arguments{
\item{sql.table}{PAS table name in sql engine}
}
\description{
Ward level time series.
}
\details{
Steps:
1. Make data set on admissions date
2. Make data set on discharge date
3. Add beds to each
4. Group data sets by adet and calculate count per mean(Beds)
}
