% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_work_demographics.R
\name{make.annual.demographics}
\alias{make.annual.demographics}
\title{Produce the annual lagged demographics data set having
reduced demographic cardinality.}
\usage{
make.annual.demographics()
}
\description{
Produce the annual lagged demographics data set having
reduced demographic cardinality.
}
