% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate_assignment_shift_aggregate.R
\name{make.assignment.shift.aggregate}
\alias{make.assignment.shift.aggregate}
\title{Make assignment shift data}
\usage{
make.assignment.shift.aggregate(
  sql.table = "JPUH_Allocate_Assignment_Combined_tsv"
)
}
\arguments{
\item{sql.table}{Name of data set in database}
}
\description{
Steps:
1. Filter data to only wards with an allocate mapping
2. Calculate diff between actual and planned staffing/ 
   embed shift over/under staffing as one hot vec.
3. Calculate sum of staffing levels by ward, date and shift
4. Calculate over/under/mixed status [see status.f for logic]
5. Save data to file
}
