% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_work_demographics.R
\name{make.demographics}
\alias{make.demographics}
\title{Produce the monthly lagged demographics data set having reduced demographic cardinality.}
\usage{
make.demographics()
}
\description{
Produce the monthly lagged demographics data set having reduced demographic cardinality.
}
