% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/establishment.R
\name{make.establishment.lag3}
\alias{make.establishment.lag3}
\title{Calculate establishment at lag 3}
\usage{
make.establishment.lag3(years = 2015:2020)
}
\arguments{
\item{years}{vector of years to be querried across}
}
\description{
Calculate establishment at lag 3
}
