% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incident.R
\name{make.incident.data}
\alias{make.incident.data}
\title{Querry incident data}
\usage{
make.incident.data(sql.table = "jpuh_ulysses_IncidentFile")
}
\arguments{
\item{sql.table}{Name of data set}
}
\description{
Steps:
1. Query data set
2. Define harmful and non-harmful incidents
3. Lag data and calculate monthly harmful/ non-harmful events per ward
4. Convert incidents to incidents per patient day
}
