% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly_assignment_metrics.R
\name{make.monthly.assignment}
\alias{make.monthly.assignment}
\title{Steps:
1. Limit analysis to shifts with at least one person in post.
2. Add bed sizes
3. Calculate shift staffing (Count and Hours) per bed
4. Calculate aggregate metrics for staffing and staff status columns.}
\usage{
make.monthly.assignment(fn = pkg.env$default_functions)
}
\arguments{
\item{fn}{List of aggregate metrics}
}
\description{
Steps:
1. Limit analysis to shifts with at least one person in post.
2. Add bed sizes
3. Calculate shift staffing (Count and Hours) per bed
4. Calculate aggregate metrics for staffing and staff status columns.
}
