% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sickness.R
\name{make.monthly.sickness}
\alias{make.monthly.sickness}
\title{Produce time lost to sickness features}
\usage{
make.monthly.sickness(sql_table = "jpuh_ESR_Sickness_processed")
}
\arguments{
\item{sql_table}{Sql database name.}
}
\description{
Convert data set into ward-month sickness fte lost reports.
}
