% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shifts_worked.R
\name{make.shifts.worked}
\alias{make.shifts.worked}
\title{Generate the shifts worked data sets (No Demographics)}
\usage{
make.shifts.worked()
}
\description{
Generate the shifts worked data sets (No Demographics)
}
