% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Leavers.R
\name{make.voluntary.leavers}
\alias{make.voluntary.leavers}
\title{Steps
1. Limit data to only wards in allocate data set
2. Calculate}
\usage{
make.voluntary.leavers(sql_table = LEAVERS_SQL_TABLE)
}
\arguments{
\item{sql_table}{Name of data set in database}
}
\description{
Steps
1. Limit data to only wards in allocate data set
2. Calculate
}
