% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incident.R
\name{make_incident_data}
\alias{make_incident_data}
\title{Querry incident data}
\usage{
make_incident_data(sql_table = "jpuh_ulysses_IncidentFile", .con = pkg_env$con)
}
\arguments{
\item{sql_table}{Name of data set}
}
\description{
Steps:
1. Query data set
2. Define harmful and non-harmful incidents
3. Lag data and calculate monthly harmful/ non-harmful events per ward
4. Convert incidents to incidents per patient day
}
