% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Acuity.Data.R
\name{make_monthly_acuity}
\alias{make_monthly_acuity}
\title{Monthly Acuity data summary}
\usage{
make_monthly_acuity(
  sql_table = "jpuh_AllocateAcuity",
  fn = pkg_env$default_functions
)
}
\arguments{
\item{sql_table}{Name of relevant sql table}

\item{fn}{List of functions to use for summarizing data}
}
\description{
Steps:
1. Get rows of data where `Name` like "Level"
2. Pivot data to ward-date vs Acuity
3. Exclude rows where no data was entered
4. Fill in missing values from remaining cases with zero.
5. Add bed sizes.
6. Calculate Acuity frequencies per bed and Acuity Score (SNCT multipliers)
7. Calculate aggregate metrics
  (Defaults: mean, var, Na ratio) at lag periods
8. Save to file
}
