% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{merge.lists.by.name}
\alias{merge.lists.by.name}
\title{Combines a series of lists based on named elements}
\usage{
\method{merge}{lists.by.name}(...)
}
\arguments{
\item{...}{A series of Lists with shared names}
}
\description{
Takes a series of lists and produces a single list, combining based on
  names.
}
