% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sickness.R
\name{monthly.sickness.f2}
\alias{monthly.sickness.f2}
\title{Calculate proportion of sickness within a given time window}
\usage{
monthly.sickness.f2(df, month.start, month.end)
}
\arguments{
\item{df}{Data frame}

\item{month.start}{Start date of time window}

\item{month.end}{End date of time window}
}
\description{
Only intended for use with `ESR sickness` data set
}
