% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{nurs.join}
\alias{nurs.join}
\title{Join data sets for NuRS study.}
\usage{
nurs.join(x, y)
}
\arguments{
\item{x, y}{data frames with fields: 'Ward', 'Year' and 'Month'

:@export}
}
\description{
Join data frames based on Ward, Year, and Month
}
