% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_work_demographics.R
\name{one.hot.encode}
\alias{one.hot.encode}
\title{Convert categorical variable to encoded matrix with dimensions 'nrow' x 'cardinality'}
\usage{
one.hot.encode(data, column)
}
\arguments{
\item{data}{A data frame}

\item{column}{A column of data to encode.}
}
\description{
Convert categorical variable to encoded matrix with dimensions 'nrow' x 'cardinality'
}
