% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_to_data.R
\name{pas_to_allocate}
\alias{pas_to_allocate}
\title{Map PAS locations to Allocate Wards}
\usage{
pas_to_allocate(values)
}
\arguments{
\item{values}{A vector of PAS locations}
}
\description{
Map PAS locations to Allocate Wards
}
