% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_per_mean.R
\name{planned.between.shift.var.per.mean.loop}
\alias{planned.between.shift.var.per.mean.loop}
\title{Convert all between shift planned staffing variances}
\usage{
planned.between.shift.var.per.mean.loop(data, grid)
}
\arguments{
\item{data}{A data frame}

\item{grid}{A frame with headings `Window`, `Staff`, `Type`.}
}
\description{
Converts from varaiance to variance per mean.
}
