% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_per_mean.R
\name{planned.between.shift.var.per.mean}
\alias{planned.between.shift.var.per.mean}
\title{Convert between shift variance for planned staffing levels.}
\usage{
planned.between.shift.var.per.mean(
  data,
  window,
  staff = "Registered",
  type = "Count"
)
}
\arguments{
\item{data}{Data frame}

\item{window}{A lagging period present in 'data'}

\item{staff}{A staff type in 'data'}

\item{type}{A staffing metric in 'data'}
}
\description{
Convert between shift variance for planned staffing levels.
}
