% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_work_demographics.R
\name{process.swd}
\alias{process.swd}
\title{Convert demographics in the shifts worked data set based on 'DemographicAnalysisGroups.xlsx'}
\usage{
process.swd(frm)
}
\arguments{
\item{frm}{A data frame of shifts worked demographics}
}
\description{
Convert demographics in the shifts worked data set based on 'DemographicAnalysisGroups.xlsx'
}
