% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{process_across_f}
\alias{process_across_f}
\title{Function factory for lagging data}
\usage{
process_across_f(data, date.col, .cols = everything(), .fn = default.fn)
}
\arguments{
\item{data}{A data frame to be operated on}

\item{date.col}{A field of 'data' that will be lagged}

\item{.cols}{The columns of 'data' to be summarized}

\item{.fn}{The summary functions to be applied [named list]}
}
\value{
A function expecting a laging time window (months) as an argument.
}
\description{
Produce a template function designed for applying a lagging transform
at different windows
}
