% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate_shift_worked_functions.R
\name{shift_worked_table}
\alias{shift_worked_table}
\title{Get sql table string for  shifts worked table by year}
\usage{
shift_worked_table(.year)
}
\arguments{
\item{.year}{Year of data to be queried}
}
\description{
Get sql table string for  shifts worked table by year
}
