% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sickness.R
\name{sickness.windows}
\alias{sickness.windows}
\title{Produce month-ward sickenss data}
\usage{
sickness.windows(data, month.starts, month.ends)
}
\arguments{
\item{data}{A data frame containing `Absence Start Date` and
`Absence End Date` fields}

\item{month.starts}{A vector of month start dates.}

\item{month.ends}{A vector of month end dates.}
}
\description{
Run `monthly.sickness.f2` across pairs of month-start/-end values
}
