% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate_assignment_shift_aggregate.R
\name{status_f}
\alias{status_f}
\title{Determine shift status.}
\usage{
status_f(delta_reg, delta_unreg, delta_other)
}
\arguments{
\item{deltra_reg}{Change in staffing from planned for registered staff}

\item{deltra_runeg}{Change in staffing from planned for unregistered staff}

\item{deltra_other}{Change in staffing from planned for other staff}
}
\description{
Rules for deciding if a shift had 'Over', 'Under', 'Fully', or 'Mixed'
staffing
}
