% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shifts_worked.R
\name{sum.to.proportion}
\alias{sum.to.proportion}
\title{Convert total work to proportion of work done by each staff group.}
\usage{
\method{sum}{to.proportion}(data)
}
\arguments{
\item{data}{A data frame}
}
\description{
Convert total work to proportion of work done by each staff group.
}
