% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shifts_worked.R
\name{sw.daily.f}
\alias{sw.daily.f}
\title{Query and convert shifts worked data set to a report of quantity of work
done each day by each staff group}
\usage{
sw.daily.f(year)
}
\arguments{
\item{year}{Numeric year of data to look up}
}
\description{
Query and convert shifts worked data set to a report of quantity of work
done each day by each staff group
}
