% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate_shift_worked_functions.R
\name{sw_to_date}
\alias{sw_to_date}
\title{Convert char to date}
\usage{
sw_to_date(values)
}
\arguments{
\item{value}{Vector of character values}
}
\description{
Cast data to date [specifically designed
  for use with shifts worked data set].
}
