% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_work_demographics.R
\name{to.chunks}
\alias{to.chunks}
\title{Subdivide a data frame into a list of data frames for divide and conquer.}
\usage{
to.chunks(data, chunk.size = 10000)
}
\arguments{
\item{data}{A data frame}

\item{chunk.size}{The max number of rows per item in list.}
}
\description{
Subdivide a data frame into a list of data frames for divide and conquer.
}
