% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_to_data.R
\name{ulys_to_allocate}
\alias{ulys_to_allocate}
\title{Map Ulysses locations to Allocate Wards}
\usage{
ulys_to_allocate(values)
}
\arguments{
\item{values}{A vector of Ulysses locations}
}
\description{
Map Ulysses locations to Allocate Wards
}
