% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_to_data.R
\docType{data}
\name{ulys_to_allocate_list}
\alias{ulys_to_allocate_list}
\title{Mappings to Allocate Ward names
Tag:Lookup
Tag:Allocate
Tag:Ulysses
Tag:ESR
Tag:PAS}
\format{
An object of class \code{list} of length 35.
}
\usage{
ulys_to_allocate_list
}
\description{
Mappings to Allocate Ward names
Tag:Lookup
Tag:Allocate
Tag:Ulysses
Tag:ESR
Tag:PAS
}
\keyword{datasets}
