% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.functions.R
\name{var.functions}
\alias{var.functions}
\title{Calculate staffing variance}
\usage{
var.functions(data)
}
\arguments{
\item{data}{A data frame with fields: 'Ward', 'Year', 'Month' and 'Shift Type'}
}
\description{
Calculate:
1. Variance between average staffing of the shifts 
2. Variance within shifts of staffing.
}
