% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_per_mean.R
\name{acuity.var.features.f}
\alias{acuity.var.features.f}
\title{Convert all raw staffing variances}
\usage{
acuity.var.features.f(data, window, feature)
}
\arguments{
\item{data}{}

\item{grid}{}
}
\description{
Converts from varaiance to variance per mean.
}
