% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate_shift_worked_functions.R,
%   R/shifts_worked.R
\name{fetch_sw}
\alias{fetch_sw}
\title{Get shifts worked data by year}
\usage{
fetch_sw(year)

fetch_sw(year)
}
\arguments{
\item{year}{Numeric year of data to querry}

\item{.year}{Year of data to be queried}

\item{.con}{SQL engine connection}
}
\description{
Get shifts worked data by year

Query and clean Shifts Worked data
}
