% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Leavers.R
\name{make.all.leavers}
\alias{make.all.leavers}
\title{Make data set of all Leavers}
\usage{
make.all.leavers(sql_table = LEAVERS_SQL_TABLE)
}
\arguments{
\item{sql_table}{Name of data set in database}
}
\description{
Make data set of all Leavers
}
